mooved the admin product from user to admin pages
